+++ Bermuda/Bermuda/GameTimer.cpp
@@ -35,8 +35,9 @@ void GameTimer::init()
 
 void GameTimer::updateGameTime(double frameTime)
 {
+
 	this->frameTime = frameTime;
-	this->gameTime += frameTime;
+	this->gameTime += static_cast<float>(frameTime);
 	this->updateDay();
 }
