+++ Bermuda/Bermuda/ImageLoader.h
@@ -38,7 +38,7 @@ private:
 public:
 	ImageLoader(SDL_Renderer* renderer);
 	~ImageLoader();
-	int loadTileset(string filename, double tileWidth, double tileHeight);
+	int loadTileset(string filename, int tileWidth, int tileHeight);
 	int getCurrentImageCount();
 
 	//Interact Green/Red
